+++ src/shapes/ellipse.js
@@ -42,8 +42,8 @@ class Ellipse {
   }
 
   scale(scalar) {
-    this.vars.width *= scalar;
-    this.vars.height *= scalar;
+    this.scaleSizeable(scalar);
+    this.scaleStyleable(scalar);
     return this;
   }