+++ test/shared/shapes/ellipse.js
@@ -71,11 +71,13 @@ describe("Rune.Ellipse", function() {
   describe("scale()", function() {
 
     it("scales the ellipse", function() {
+      spyOn(s, 'scaleSizeable');
+      spyOn(s, 'scaleStyleable');
       s.scale(2);
       expect(s.vars.x).toEqual(10);
       expect(s.vars.y).toEqual(15);
-      expect(s.vars.width).toEqual(600);
-      expect(s.vars.height).toEqual(610);
+      expect(s.scaleSizeable).toHaveBeenCalledWith(2);
+      expect(s.scaleStyleable).toHaveBeenCalledWith(2);
     });
 
   });