+++ test/shared/shapes/polygon.js
@@ -208,6 +208,7 @@ describe("Rune.Polygon", function() {
   describe("scale()", function() {
 
     it("scales the polygon", function() {
+      spyOn(s, 'scaleStyleable');
       s.scale(2);
       expect(s.vars.x).toEqual(10);
       expect(s.vars.y).toEqual(15);
@@ -215,6 +216,7 @@ describe("Rune.Polygon", function() {
       expect(s.vars.vectors[1]).toEqualVector(120, 0);
       expect(s.vars.vectors[2]).toEqualVector(160, 120);
       expect(s.vars.vectors[3]).toEqualVector(40, 120);
+      expect(s.scaleStyleable).toHaveBeenCalledWith(2);
     });
 
   });