+++ test/shared/shapes/text.js
@@ -84,10 +84,12 @@ describe("Rune.Text", function() {
   });
 
   it("scales the rectangle", function() {
+    spyOn(s, 'scaleStyleable');
     s.scale(2);
     expect(s.vars.x).toEqual(10);
     expect(s.vars.y).toEqual(15);
     expect(s.vars.fontSize).toEqual(32);
+    expect(s.scaleStyleable).toHaveBeenCalledWith(2);
   });
 
 });