+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.14
 
 - Adding `Rune.Image` to draw images in a SVG element
+- `strokeWidth` is now scaled when calling `.scale()` on a shape
 
 ## 0.2.13